On the way to flipping dominos(4) : selecting dominos OK.
