+++ sage_combinat_widgets/grid_view_editor.py
@@ -502,7 +502,10 @@ class GridViewEditor(BindableEditorClass):
             raise Exception("Removing cells is not implemented for this object.")
         if val == True: # if it's a button, reverse button toggling
             val = False
-        obj = copy(self.value)
+        try:
+            obj = copy(self.value)
+        except:
+            return # If we can't copy the object, then we cannot continue anyway FIXME how to make sure copy is possible?
         result = self.adapter.remove_cell(obj, pos, dirty=self.dirty)
         if issubclass(result.__class__, BaseException): # Removing cell was impossible
             if pos in self.addable_cells() or (pos in self.cells and val == self.cells[pos]) and self.dirty.keys() == [pos]: # Rollback
